requirements.txt
asgiref==3.2.3
Django==3.0.3
django-multiselectfield==0.1.12
Pillow==7.0.0
psycopg2==2.8.4
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.0
